Split logic validation
